% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{print.m2}
\alias{print.m2}
\title{Tropical arithmetic}
\usage{
\method{print}{m2}(x, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}
}
\value{
.
}
\description{
Tropical arithmetic
}
\examples{

m2("13^1000")


}
