% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{tarithmetic}
\alias{tarithmetic}
\alias{\%+\%}
\alias{\%.\%}
\alias{set_plus_min}
\alias{set_plus_max}
\alias{get_plus_type}
\title{Tropical arithmetic}
\usage{
x \%+\% y

x \%.\% y

set_plus_min(message = TRUE)

set_plus_max(message = TRUE)

get_plus_type()
}
\arguments{
\item{x, y}{a numeric data structure (vector/matrix/array)}

\item{message}{issue messages to user? (default: TRUE)}
}
\value{
.
}
\description{
Tropical arithmetic
}
\examples{

get_plus_type()

1 \%+\% 5
1:3 \%+\% 3:1

set_plus_max()
get_plus_type()

1 \%+\% 5
1:3 \%+\% 3:1

set_plus_min()
get_plus_type()

}
